// src/pages/TestUnitTableList/data.d.ts
export interface TableListItem {
  key: number;
  disabled?: boolean;

  testUnitName: string;

  testUnitDesc: string;

  status: number;
  updatedAt: Date;
  createdAt: Date;
  progress: number;
}

export interface TableListPagination {
  total: number;
  pageSize: number;
  current: number;
}

export interface TableListData {
  list: TableListItem[];
  pagination: Partial<TableListPagination>;
}

export interface TableListParams {
  status?: string;
  testUnitName?: string;
  testUnitDesc?: string;
  key?: number;
  pageSize?: number;
  currentPage?: number;
  filter?: { [key: string]: any[] };
  sorter?: { [key: string]: any };
}